Model list cleanup: shared omni-gen hooks, dead code removal
