+++ frontend/apps/artcraft-webapp/src/components/topbar/task-queue.tsx
@@ -23,6 +23,7 @@ import { Button } from "@storyteller/ui-button";
 import {
   getProviderDisplayName,
   getModelDisplayName,
+  getCreatorIconPathForModelId,
   findModelByKey,
 } from "@storyteller/model-list";
 import { CloseButton } from "@storyteller/ui-close-button";
@@ -30,7 +31,6 @@ import dayjs from "dayjs";
 import { ActionReminderModal } from "@storyteller/ui-action-reminder-modal";
 import { Lightbox } from "../lightbox/lightbox";
 import { showToast } from "../toast/toast";
-import { getModelCreatorIconPath } from "../../lib/omni-gen-hooks";
 import { useRecreateFromPromptToken } from "../../lib/recreate";
 import { getMediaThumbnail, THUMBNAIL_SIZES } from "@storyteller/common";
 import { twMerge } from "tailwind-merge";
@@ -263,7 +263,7 @@ const InProgressCard = ({
       ? formatTimeLeft(task.estimatedTimeLeftMs)
       : null;
   const modelIconPath = task.modelType
-    ? getModelCreatorIconPath(task.modelType)
+    ? getCreatorIconPathForModelId(task.modelType)
     : null;
 
   return (
