+++ frontend/apps/artcraft-webapp/tsconfig.app.json
@@ -47,13 +47,16 @@
       "path": "../../libs/components/upload-modal/tsconfig.lib.json"
     },
     {
-      "path": "../../libs/components/moodboard/tsconfig.lib.json"
-    },
-    {
       "path": "../../libs/components/pagedraw/tsconfig.lib.json"
     },
     {
+      "path": "../../libs/components/moodboard/tsconfig.lib.json"
+    },
+    {
       "path": "../../libs/components/slider-v2/tsconfig.lib.json"
+    },
+    {
+      "path": "../../libs/omni-gen/tsconfig.lib.json"
     },
     {
       "path": "../../libs/components/pricing-modal/tsconfig.lib.json"
