+++ frontend/apps/artcraft-website/src/main.tsx
@@ -5,6 +5,8 @@ import { GoogleOAuthProvider } from "@react-oauth/google";
 import App from "./app/app";
 import { StorytellerApiHostStore, UsersApi } from "@storyteller/api";
 import { captureLandingContext, getReferrer } from "@storyteller/common";
+import { setOmniGenErrorNotifier } from "@storyteller/omni-gen";
+import { toast } from "./components/toast/toast";
 
 const GOOGLE_CLIENT_ID = import.meta.env.VITE_GOOGLE_CLIENT_ID;
 
@@ -24,6 +26,9 @@ if (import.meta.env.DEV) {
     console.warn("Failed to set dev API host override", e);
   }
 }
+
+// Surface omni model/generation outages through this app's toast component.
+setOmniGenErrorNotifier((message) => toast.error(message));
 
 // Persist landing context (referral username, landing URL, referrer) to apex-
 // domain cookies so attribution survives the getartcraft.com →
