+++ frontend/apps/artcraft-website/tsconfig.app.json
@@ -50,6 +50,9 @@
       "path": "../../libs/components/gallery-modal/tsconfig.lib.json"
     },
     {
+      "path": "../../libs/omni-gen/tsconfig.lib.json"
+    },
+    {
       "path": "../../libs/components/switch/tsconfig.lib.json"
     },
     {
@@ -65,13 +68,13 @@
       "path": "../../libs/components/gravatar/tsconfig.lib.json"
     },
     {
-      "path": "../../libs/model-list/tsconfig.lib.json"
-    },
-    {
       "path": "../../libs/components/loading-spinner/tsconfig.lib.json"
     },
     {
       "path": "../../libs/components/tooltip/tsconfig.lib.json"
+    },
+    {
+      "path": "../../libs/model-list/tsconfig.lib.json"
     },
     {
       "path": "../../libs/components/popover/tsconfig.lib.json"
