+++ frontend/libs/model-list/src/index.ts
@@ -8,6 +8,7 @@ export * from "./lib/classes/metadata/ModelCreator.js";
 export * from "./lib/classes/metadata/ModelCreatorIcons.js";
 export * from "./lib/classes/metadata/ModelCreatorIconForId.js";
 export * from "./lib/classes/metadata/ModelMapping.js";
+export * from "./lib/classes/metadata/ModelDescriptions.js";
 export * from "./lib/classes/metadata/PromptFormatters.js";
 export * from "./lib/classes/metadata/ModelTag.js";
 export * from "./lib/classes/metadata/SizeOption.js";
@@ -22,5 +23,4 @@ export * from "./lib/classes/SplatModel.js";
 export * from "./lib/lists/SplatModels.js";
 export * from "./lib/classes/Object3DModel.js";
 export * from "./lib/lists/Object3DModels.js";
-export * from "./lib/loader/modelReconciliation.js";
 export * from "./lib/loader/buildModelsFromListing.js";
