+++ frontend/libs/omni-gen/README.md
@@ -1,0 +1,7 @@
+# @storyteller/omni-gen
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test @storyteller/omni-gen` to execute the unit tests via [Vitest](https://vitest.dev/).
