+++ frontend/libs/omni-gen/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./lib/omni-gen-hooks.js";
