+++ frontend/tsconfig.json
@@ -212,6 +212,9 @@
     },
     {
       "path": "./libs/components/moodboard"
+    },
+    {
+      "path": "./libs/omni-gen"
     }
   ]
 }
